# Store MotionSubspaces in a TypeSortedCollection