--- a/src/mechanism_algorithms.jl
+++ b/src/mechanism_algorithms.jl
@@ -172,23 +172,21 @@ function mass_matrix!(out::Symmetric{C, Matrix{C}}, state::MechanismState{X, M, 
     @boundscheck out.uplo == 'L' || error("expected a lower triangular symmetric matrix type as the mass matrix")
     update_motion_subspaces_in_world!(state)
     update_crb_inertias!(state)
+    fill!(out.data, 0)
     joints = state.type_sorted_tree_joints
     foreach_with_extra_args(out, state, joints) do out, state, jointi # TODO: use closure once it doesn't allocate
-        Base.@_inline_meta
         irange = velocity_range(state, jointi)
         Si = motion_subspace_in_world(state, jointi)
         Ici = crb_inertia(state, successor(jointi, state.mechanism))
         F = Ici * Si
         ancestor_joint_ids = state.ancestor_joint_ids[jointi]
         foreach_with_extra_args(out, state, irange, F, ancestor_joint_ids, state.type_sorted_tree_joints) do out, state, irange, F, ancestor_joint_ids, jointj # TODO: use closure once it doesn't allocate
-            Base.@_inline_meta
-            jrange = velocity_range(state, jointj)
-            if Int(Graphs.edge_id(jointj)) ∈ ancestor_joint_ids
+            Base.@_inline_meta # currently required; try removing with 1.0
+            if ancestor_joint_ids[Int(Graphs.edge_id(jointj))]
+                jrange = velocity_range(state, jointj)
                 Sj = motion_subspace_in_world(state, jointj)
                 block = angular(F)' * angular(Sj) + linear(F)' * linear(Sj)
                 set_matrix_block!(out.data, irange, jrange, block)
-            else
-                zero_matrix_block!(out.data, irange, jrange)
             end
         end
     end