--- a/src/mechanism_state.jl
+++ b/src/mechanism_state.jl
@@ -19,7 +19,7 @@ struct MechanismState{X, M, C, JointCollection}
     modcount::Int
     type_sorted_tree_joints::JointCollection
     type_sorted_non_tree_joints::JointCollection
-    type_sorted_ancestor_joints::JointDict{M, JointCollection}
+    ancestor_joint_ids::JointDict{M, Vector{JointID}}
     constraint_jacobian_structure::JointDict{M, TreePath{RigidBody{M}, GenericJoint{M}}}
 
     q::Vector{X} # configurations
@@ -55,8 +55,8 @@ struct MechanismState{X, M, C, JointCollection}
         JointCollection = typeof(type_sorted_joints)
         type_sorted_tree_joints = JointCollection(typedjoint.(tree_joints(mechanism)))
         type_sorted_non_tree_joints = JointCollection(typedjoint.(non_tree_joints(mechanism)))
-        ancestor_joints(joint) = path(mechanism, successor(joint, mechanism), root_body(mechanism)).edges
-        type_sorted_ancestor_joints = JointDict{M, JointCollection}(j => JointCollection(typedjoint.(ancestor_joints(j))) for j in tree_joints(mechanism))
+        ancestor_joint_id_list = joint -> sort(Graphs.edge_id.(path(mechanism, successor(joint, mechanism), root_body(mechanism)).edges), by = Int)
+        ancestor_joint_ids = JointDict{M, Vector{JointID}}(j => ancestor_joint_id_list(j) for j in tree_joints(mechanism))
 
         # joint-specific
         qstart, vstart = 1, 1
@@ -91,7 +91,7 @@ struct MechanismState{X, M, C, JointCollection}
         m = mechanism
         constraint_jacobian_structure = JointDict{M, TreePath{RigidBody{M}, GenericJoint{M}}}(j => path(m, predecessor(j, m), successor(j, m)) for j in non_tree_joints(m))
 
-        new{X, M, C, JointCollection}(mechanism, modcount(mechanism), type_sorted_tree_joints, type_sorted_non_tree_joints, type_sorted_ancestor_joints,
+        new{X, M, C, JointCollection}(mechanism, modcount(mechanism), type_sorted_tree_joints, type_sorted_non_tree_joints, ancestor_joint_ids,
             constraint_jacobian_structure, q, v, s, qs, vs, joint_poses,
             joint_transforms, joint_twists, joint_bias_accelerations, motion_subspaces_in_world, constraint_wrench_subspaces,
             transforms_to_root, twists_wrt_world, bias_accelerations_wrt_world, inertias, crb_inertias,