--- a/src/mechanism_state.jl
+++ b/src/mechanism_state.jl
@@ -19,7 +19,7 @@ struct MechanismState{X, M, C, JointCollection}
     modcount::Int
     type_sorted_tree_joints::JointCollection
     type_sorted_non_tree_joints::JointCollection
-    ancestor_joint_ids::JointDict{M, IntSet}
+    ancestor_joint_ids::JointDict{M, BitVector}
     constraint_jacobian_structure::JointDict{M, TreePath{RigidBody{M}, GenericJoint{M}}}
 
     q::Vector{X} # configurations
@@ -55,8 +55,8 @@ struct MechanismState{X, M, C, JointCollection}
         JointCollection = typeof(type_sorted_joints)
         type_sorted_tree_joints = JointCollection(typedjoint.(tree_joints(mechanism)))
         type_sorted_non_tree_joints = JointCollection(typedjoint.(non_tree_joints(mechanism)))
-        ancestor_joint_id_set = joint -> IntSet(Int.(Graphs.edge_id.(path(mechanism, successor(joint, mechanism), root_body(mechanism)).edges)))
-        ancestor_joint_ids = JointDict{M, IntSet}(j => ancestor_joint_id_set(j) for j in tree_joints(mechanism))
+        ancestor_joint_id_bitvec = joint -> BitArray(j ∈ path(mechanism, successor(joint, mechanism), root_body(mechanism)) for j in tree_joints(mechanism))
+        ancestor_joint_ids = JointDict{M, BitVector}(j => ancestor_joint_id_bitvec(j) for j in tree_joints(mechanism))
 
         # joint-specific
         qstart, vstart = 1, 1