Previously, motion subspaces were converted to a common type (a `GeometricJacobian` backed by 3x6 `SMatrix`es) so that they could be stored type-stably in `MechanismState`. This had some overhead associated with it, but this is just what `TypeSortedCollection`s are made for.

This change required no longer storing ancestor joints for all of the tree joints in a `TypeSortedCollection` (`state.type_sorted_tree_joints`), as the indices of the `TypeSortedCollection` we iterate over in the inner loop of `mass_matrix!` need to match those of the `TypeSortedCollection` of `MotionSubspace`s. Instead, I'm now using a `JointMask`, which is a little wrapper around a `BitVector`.

Decent performance improvements:

* `momentum_matrix!`: 9.146 μs to 7.081 μs
* `dynamics!`: 47.448 μs to 43.964 μs
* `geometric_jacobian!`: 5.669 μs to 4.073 μs
* `mass_matrix!`: 13.727 μs to 12.286 μs

`constraint_jacobian_and_bias!` went from `99.836 μs` to `109.052 μs`, but I haven't put any effort into optimizing that one anyway.

Unfortunately, this means that the `motion_subspace_in_world` function has to go away (or be extremely slow and annoying to implement), but I don't expect any users to be using that function. Please open an issue if you do.